+++ vowpalwabbit/io_buf.h
@@ -221,7 +221,7 @@ inline size_t bin_read_fixed(io_buf& i, char* data, size_t len, const char* read
 
     // compute hash for check-sum
     if (i.verify_hash)
-      i.hash = uniform_hash(p, len, i.hash);
+      i.hash = (uint32_t)uniform_hash(p, len, i.hash);
 
     if (*read_message == '\0')
       memcpy(data,p,len);
@@ -251,7 +251,7 @@ inline size_t bin_write_fixed(io_buf& o, const char* data, uint32_t len)
 
     // compute hash for check-sum
     if (o.verify_hash)
-    { o.hash = uniform_hash(p, len, o.hash);
+    { o.hash = (uint32_t)uniform_hash(p, len, o.hash);
     }
   }
   return len;
